Allow show_chunks to process more than 65535 chunks
(cherry picked from commit 18d492ee1ac5dbf66f20e369f45f1eb13f750d3e)

+++ src/chunk.h
@@ -102,7 +102,7 @@ typedef struct ChunkScanCtx
 	const Hypertable *ht;
 	const Point *point;
 	unsigned int num_complete_chunks;
-	int num_processed;
+	uint64 num_processed;
 	bool early_abort;
 	LOCKMODE lockmode;
